Fix (U)Int 64 reading

+++ src/ByteArray/Types/UInt64.cs
@@ -25,24 +25,24 @@
             if (endianness == Endianness.BE)
             {
                 result = this.buffer[position + 7];
-                result |= (uint)(this.buffer[position + 6] << 8);
-                result |= (uint)(this.buffer[position + 5] << 16);
-                result |= (uint)(this.buffer[position + 4] << 24);
-                result |= (uint)(this.buffer[position + 3] << 32);
-                result |= (uint)(this.buffer[position + 2] << 40);
-                result |= (uint)(this.buffer[position + 1] << 48);
-                result |= (uint)(this.buffer[position + 0] << 56);
+                result |= (ulong)this.buffer[position + 6] << 8;
+                result |= (ulong)this.buffer[position + 5] << 16;
+                result |= (ulong)this.buffer[position + 4] << 24;
+                result |= (ulong)this.buffer[position + 3] << 32;
+                result |= (ulong)this.buffer[position + 2] << 40;
+                result |= (ulong)this.buffer[position + 1] << 48;
+                result |= (ulong)this.buffer[position + 0] << 56;
             }
             else
             {
                 result = this.buffer[position + 0];
-                result |= (uint)(this.buffer[position + 1] << 8);
-                result |= (uint)(this.buffer[position + 2] << 16);
-                result |= (uint)(this.buffer[position + 3] << 24);
-                result |= (uint)(this.buffer[position + 4] << 32);
-                result |= (uint)(this.buffer[position + 5] << 40);
-                result |= (uint)(this.buffer[position + 6] << 48);
-                result |= (uint)(this.buffer[position + 7] << 56);
+                result |= (ulong)this.buffer[position + 1] << 8;
+                result |= (ulong)this.buffer[position + 2] << 16;
+                result |= (ulong)this.buffer[position + 3] << 24;
+                result |= (ulong)this.buffer[position + 4] << 32;
+                result |= (ulong)this.buffer[position + 5] << 40;
+                result |= (ulong)this.buffer[position + 6] << 48;
+                result |= (ulong)this.buffer[position + 7] << 56;
             }
             return result;
         }
